ponderada_03/back/logger/requirements.txt
flask[async]
Flask-SQLAlchemy
Flask-JWT-Extended
requests